add computer module
